Set to editor type app
Changed app type to editor for SDK 14.0.0.
