+++ viktor.config.toml
@@ -1,2 +1,2 @@
-app_type = 'tree'
+app_type = 'editor'
 python_version = '3.11'  # '3.8' | '3.9' | '3.10' | '3.11'
